+++ CHANGELOG.md
@@ -20,6 +20,7 @@ We haven't had the best track record of code/feature changes before this date, b
     </summary>
 
   - [bug] Fix IntlNumberType.get(...) to work with short locales https://github.com/facebookincubator/fbt/pull/85
+  - Allow fbt:enum value prop to be a string literal
 
   </details>
 - 0.10.2: